pyproject.toml
[project]
name = "zizevenmore"
version = "0.1.0"
description = ""
authors = [
    {name = "David McKee",email = "david.mckee@dxw.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "requests-cache (>=1.2.1,<2.0.0)",
    "pygithub (>=2.6.1,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
